backend/requirements.txt
# Web Framework
Flask==3.1.0
flask-cors==5.0.0
Werkzeug==3.1.5
blinker==1.9.0

# Database & Migrations
SQLAlchemy==2.0.46
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.10
alembic==1.18.3

# Auth
Flask-JWT-Extended==4.7.1
bcrypt==4.2.1
PyJWT==2.11.0

# Services (Mail & Images)
flask-mailman==1.1.0
email-validator==2.2.0
Pillow==11.1.0

# Environment & Utils
python-dotenv==1.0.1
python-dateutil==2.9.0
typing_extensions==4.12.2

# Production Server
gunicorn==23.0.0